Move metrics tracker into shared reporting utilities
- relocate src/plugin/src/reporting/metrics-tracker.js to src/shared/reporting/metrics-tracker.js so the reusable metrics helper lives alongside shared utilities instead of being scoped to plugin reporting

- update imports, add a shared reporting barrel, and move the metrics tracker unit test into src/shared/tests
