+++ src/shared/reporting/index.js
@@ -1,0 +1,1 @@
+export * from "./metrics-tracker.js";
